All tests now passing
added a line inside of the decrement function to exit the function if
the previous value on the session time state is below 1. Somehow the
test suite was decrementing the value of state whilst also not updating
the state itself. Not an issue in any other test that decrements and
app is fully functional.